Reorganize new code
