+++ captum/optim/_utils/circuits.py
@@ -4,47 +4,8 @@ from typing import List, Union
 import torch
 import torch.nn as nn
 
-from captum.optim._core.output_hook import AbortForwardException, ModuleOutputsHook
 from captum.optim._utils.typing import ModuleOutputMapping
-
-
-def max2avg_pool(model) -> None:
-    """
-    Convert MaxPool2d layers to their AvgPool2d equivalents.
-    """
-
-    for name, child in model._modules.items():
-        if isinstance(child, nn.MaxPool2d):
-            new_layer = nn.AvgPool2d(
-                kernel_size=child.kernel_size,
-                stride=child.stride,
-                padding=child.padding,
-                ceil_mode=child.ceil_mode,
-            )
-            setattr(model, name, new_layer)
-        elif child is not None:
-            max2avg_pool(child)
-
-
-class ActivationCatcher(object):
-    """
-    Simple module for collecting activations from model targets.
-    """
-
-    def __init__(self, targets: Union[nn.Module, List[nn.Module]]) -> None:
-        super(ActivationCatcher, self).__init__()
-        self.layers = ModuleOutputsHook(targets)
-
-    def __call__(self, model, input_t: torch.Tensor) -> ModuleOutputMapping:
-        try:
-            with suppress(AbortForwardException):
-                model(input_t)
-            activations = self.layers.consume_outputs()
-            self.layers.remove_hooks()
-            return activations
-        except (Exception, BaseException) as e:
-            self.layers.remove_hooks()
-            raise e
+from captum.optim._utils.circuits import ActivationCatcher
 
 
 def get_expanded_weights(model, target1: nn.Module, target2: nn.Module) -> torch.Tensor:
