+++ captum/optim/_utils/models.py
@@ -5,6 +5,9 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
+from captum.optim._core.output_hook import AbortForwardException, ModuleOutputsHook
+from captum.optim._utils.typing import ModuleOutputMapping
+
 
 def get_model_layers(model) -> List[str]:
     """
@@ -85,6 +88,24 @@ def replace_layers(model, old_layer=ReluLayer, new_layer=RedirectedReluLayer) ->
             replace_layers(child, old_layer, new_layer)
 
 
+def max2avg_pool(model) -> None:
+    """
+    Convert MaxPool2d layers to their AvgPool2d equivalents.
+    """
+
+    for name, child in model._modules.items():
+        if isinstance(child, nn.MaxPool2d):
+            new_layer = nn.AvgPool2d(
+                kernel_size=child.kernel_size,
+                stride=child.stride,
+                padding=child.padding,
+                ceil_mode=child.ceil_mode,
+            )
+            setattr(model, name, new_layer)
+        elif child is not None:
+            max2avg_pool(child)
+
+
 class LocalResponseNormLayer(nn.Module):
     """
     Basic Hookable Local Response Norm layer.
@@ -151,3 +172,24 @@ class Conv2dSame(nn.Conv2d):
             self.dilation,
             self.groups,
         )
+
+
+class ActivationCatcher(object):
+    """
+    Simple module for collecting activations from model targets.
+    """
+
+    def __init__(self, targets: Union[nn.Module, List[nn.Module]]) -> None:
+        super(ActivationCatcher, self).__init__()
+        self.layers = ModuleOutputsHook(targets)
+
+    def __call__(self, model, input_t: torch.Tensor) -> ModuleOutputMapping:
+        try:
+            with suppress(AbortForwardException):
+                model(input_t)
+            activations = self.layers.consume_outputs()
+            self.layers.remove_hooks()
+            return activations
+        except (Exception, BaseException) as e:
+            self.layers.remove_hooks()
+            raise e
